# src/main/resources/openapi.yml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /devices:
    put:
      tags:
        - device-controller
      operationId: updateDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdateRecord"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeviceUpdateRecord"
    post:
      tags:
        - device-controller
      operationId: createDevice
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceRecord"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeviceRecord"
  /devices/{id}:
    get:
      tags:
        - device-controller
      operationId: getDeviceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DeviceRecord"
    delete:
      tags:
        - device-controller
      operationId: deleteDeviceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /devices/all:
    get:
      tags:
        - device-controller
      operationId: getAllDevices
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceRecord"
  /devices/all-by-state:
    get:
      tags:
        - device-controller
      operationId: getAllDevicesByState
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: string
            enum:
              - AVAILABLE
              - IN_USE
              - INACTIVE
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceRecord"
  /devices/all-by-brand:
    get:
      tags:
        - device-controller
      operationId: getAllDevicesByBrand
      parameters:
        - name: brand
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceRecord"
components:
  schemas:
    DeviceUpdateRecord:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        brand:
          type: string
        state:
          type: string
          enum:
            - AVAILABLE
            - IN_USE
            - INACTIVE
    DeviceRecord:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        brand:
          type: string
        state:
          type: string
          enum:
            - AVAILABLE
            - IN_USE
            - INACTIVE
        creationTime:
          type: string
          format: date